Move EOL handing in from_text().
Previously, the rdata subclass from_text() routine was expected to
consume the EOL (and not all of them did).  After this change, the
common code parses the EOL token, which means that if the subclass
is parsing a variable number of tokens, it needs to unget the EOL token.

This adds a new helper method to the tokenizer, get_remaining(), to get
the remaining tokens on the line, with an optional maximum, and converts
a number of types to it.  This means that no types need to deal with
ungetting the final token explicitly, as they're all using helpers.